+++ docker/php/fpm/bin/docker-entrypoint.sh
@@ -11,10 +11,10 @@ if [ "$1" = 'php-fpm' ] || [ "$1" = 'php' ] || [ "$1" = 'bin/console' ]; then
 	# but a `vendor` directory does not exists,
 	# install required packages.
 	# Framework usage like Symfony etc.
-	if [ -f composer.json ] && [ ! -d vendor ]; then
-		echo "Install vendor packages required!"
-		composer install --prefer-dist --no-progress --no-interaction
-	fi
+	#if [ -f composer.json ] && [ ! -d vendor ]; then
+	#	echo "Install vendor packages required!"
+	#	composer install --prefer-dist --no-progress --no-interaction
+	#fi
 
 	# Correct file/directory permissions
 	setfacl -R -m u:www-data:rwX -m u:"$(whoami)":rwX /srv/app